bridge/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "fancontrolbridge"
version = "0.0.1"
authors = [
  { name="Markus Ijäs", email="markus.ijas@pm.me" },
]
description = "A MQTT-enabled bridge for the Arduino fan controller."
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GPL v3 License",
    "Operating System :: OS Independent",
]
dependencies = ["confuse", "paho-mqtt", "pyserial"]

[project.urls]
"Homepage" = "https://github.com/mtijas/mqtt-fancontroller"
"Bug Tracker" = "https://github.com/mtijas/mqtt-fancontroller/issues"

[project.scripts]
fancontrolbridge = "fancontrolbridge.fancontrolbridge:main"
